refactor: Update package structure and fix context references for MainActivity
