+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         android:theme="@style/Theme.CareX"
         tools:targetApi="31">
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
